feat: #1210: remove unused files
